fp16 frozen embedding + CF-cacheable layer-sliced packages + dl. CDN URLs
- export: save frozen emb_weight as fp16 (662MB->331MB), halving the repeated
  per-round volunteer download and fitting it under Cloudflare Free's 512MB
  cache limit. Worker copy_ casts fp16->fp32 on load (frozen tensor, amp=bf16).
- side_cycle: GPU/public export now uses a rotating 3-of-7-layer window
  (~393MB block package < 512MB CF cache), advancing the offset by 3 each round
  so all layers train over ~3 rounds (DiffusionBlocks layer-sharing).
- network_host: serve package/frozen downloads from a CF-fronted dl. subdomain
  (cached) while keeping the API + large result-upload on join. (unproxied,
  bypasses CF's 100MB upload-body limit). Configurable via AGILLM_DL_BASE.

+++ agillm4/training_bench/agillm4_export_bench_packages.py
@@ -134,7 +134,10 @@ def main() -> int:
         "tie_weights": tie_weights,
         "tokenizer_id": ck.get("tokenizer_id"),
         "vocab": vocab,
-        "emb_weight": core["emb.weight"].detach().cpu(),
+        # AGILLM_FP16_FROZEN: fp16 the frozen embedding (662MB->331MB): halves the
+        # repeated per-round volunteer download and drops it under CF Free 512MB cache.
+        # Frozen (not trained); worker copy_ casts back to fp32, amp uses bf16 anyway.
+        "emb_weight": core["emb.weight"].detach().cpu().to(torch.float16),
         "ln_weight": core["ln.weight"].detach().cpu(),
         "ln_bias": core["ln.bias"].detach().cpu(),
     }
